scripts/article_scraper/requirements.txt
selenium==4.20.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
lxml==5.2.1
PyYAML==6.0.1
rich==13.7.1
requests==2.31.0
kafka-python==2.0.2
pymongo==4.10.1
